+++ tests/test_jsonnet_loader.py
@@ -1,10 +1,9 @@
-from firecore import jsonnet_loader
 import json
-
+import rjsonnet
 
 def test_eval_snippet():
     snippet = "{a:1, b:2}"
     expected = dict(a=1, b=2)
-    output = jsonnet_loader.evaluate_snippet("snippet", snippet)
+    output = rjsonnet.evaluate_snippet("snippet", snippet)
     output = json.loads(output)
     assert output == expected